fix: bug determining rel path in filerevs command
The filrevs command was only using the included
branchs/tags to determine the relative path of the target files.
If the working copy branch was not in the included branches then
the we would fail to determine the relative path to the target file.
Now we always include the working copy branch when determinining the
relative path.